Modify arch files and config for CUDA builds
Split arch files for CUDA build into a base file ${NAMD_ARCH}.cuda
and an include file for the CUDA major version ${NAMD_ARCH}.cuda#.
Splitting the arch files removes a lot of redundancy in the files.
The major version support file is included second in Make.config,
so can override the base file settings in the event that something
fundamental changes in future releases of CUDA. Both files must be
included from Make.config. Maintain support for CUDA 8 through 11.
Remove support for CUDA 7 (except for MacOSX).

Add config option "--with-cuda-profiling" for CUDA profiling with NVTX,
which then modifies flags by including ${NAMD_ARCH}.cudaprof file.
Fixed the .cudaprof files.

If "--cuda-prefix" is not set, config attempts to determine path to
the CUDA installation by checking to see if nvcc is in the user path.
There are additional warnings emitted by config if no arch files are
found for the major version or if expected library files are not found.

Updates included for Linux-x86_64, Linux-POWER, Linux-ARM64, Win64,
and MacOSX-x86_64. Since cudart and cufft libraries are statically
linked for Linux and Mac targets, avoid copying the shared library by
setting LIBCUDARTSO and LIBCUFFTSO to be empty.

Change-Id: I4e28d82e797614ef6791e00d7ec369822e7f455a