Fix on pairlist bloating for Nbond GPU Kernel
After the previous change on the pairlist generation mechanism
for the nonbonded kernel
(https://charm.cs.illinois.edu/gerrit/c/namd/+/5298), we missed
enforcing the call to CudaTileListKernel::reSortTileLists(), which
was still following the atom migration rule.

This caused pairlists to be lean on the start of a migration cycle,
but bloated mid-cycle.

Fix by ensuring a call to reSortTileList().

Change-Id: I7c40005a0f8f02582aac73a0f8c3cd000830c981