Fix reading Drude parameters in MPI build.
Commit 9cb2fcf7647eb076b2a6402b497a951c9ddc7715 added the Thole
support for Drude force field, but it was not serialized via
Molecule::send_Molecule. Also, Commit f300bda91efcef80ee5d0f1ca86ac4f93e910365
added a variable named "numZeroMassAtoms" and used the value of it
to determine the number of lone pairs, but it was not serialized,
either. As a result, the MPI build is broken for simulating with
Drude force field.

This commit fixes the parsing of Drude parameters in the MPI build
by serializing "numTholes", "tholes" and "numZeroMassAtoms" in
the master node, and reading these variables in the client nodes.

This issue was initially reported by Selemon in the mailing list:
https://www.ks.uiuc.edu/Research/namd/mailing_list/namd-l.2020-2021/0436.html

Change-Id: If3c973a51bb8bc1e73433800f9dbec92b112b292