Fix doc of the "callback" TCL command.
The callback function is actually called after the finish of a "run"
procedure, not every step. This commit clarifies that in the manual,
and provides an example of using callback. The writeup for callback
is moved after the discussion of "run" and "minimize," and the example
clarifies use of all three commands.

Change-Id: I0adadf80ac42dc8992c077bd9d940ba8ae181c88